+++ API/DTOs/PowerPlant.cs
@@ -4,7 +4,7 @@ namespace API.DTOs
     {
         public string Name { get; set; }
         public string Type { get; set; }
-        public double Efficiency { get; set; }
+        public decimal Efficiency { get; set; }
         public int Pmin { get; set; }
         public int Pmax { get; set; }
     }
